Create CodegenCppVisitor and use common contructor for CodegenCoreneuronCppVisitor and CodegenNeuronCppVisitor
