+++ src/codegen/codegen_coreneuron_cpp_visitor.hpp
@@ -24,6 +24,7 @@
 #include <string_view>
 #include <utility>
 
+#include "codegen/codegen_cpp_visitor.hpp"
 #include "codegen/codegen_info.hpp"
 #include "codegen/codegen_naming.hpp"
 #include "printer/code_printer.hpp"
@@ -36,125 +37,6 @@
 namespace nmodl {
 
 namespace codegen {
-
-/**
- * \defgroup codegen Code Generation Implementation
- * \brief Implementations of code generation backends
- *
- * \defgroup codegen_details Codegen Helpers
- * \ingroup codegen
- * \brief Helper routines/types for code generation
- * \{
- */
-
-/**
- * \enum BlockType
- * \brief Helper to represent various block types
- *
- * Note: do not assign integers to these enums
- *
- */
-enum class BlockType {
-    /// initial block
-    Initial,
-
-    /// constructor block
-    Constructor,
-
-    /// destructor block
-    Destructor,
-
-    /// breakpoint block
-    Equation,
-
-    /// derivative block
-    State,
-
-    /// watch block
-    Watch,
-
-    /// net_receive block
-    NetReceive,
-
-    /// before / after block
-    BeforeAfter,
-
-    /// fake ending block type for loops on the enums. Keep it at the end
-    BlockTypeEnd
-};
-
-
-/**
- * \enum MemberType
- * \brief Helper to represent various variables types
- *
- */
-enum class MemberType {
-    /// index / int variables
-    index,
-
-    /// range / double variables
-    range,
-
-    /// global variables
-    global,
-
-    /// thread variables
-    thread
-};
-
-
-/**
- * \class IndexVariableInfo
- * \brief Helper to represent information about index/int variables
- *
- */
-struct IndexVariableInfo {
-    /// symbol for the variable
-    const std::shared_ptr<symtab::Symbol> symbol;
-
-    /// if variable resides in vdata field of NrnThread
-    /// typically true for bbcore pointer
-    bool is_vdata = false;
-
-    /// if this is pure index (e.g. style_ion) variables is directly
-    /// index and shouldn't be printed with data/vdata
-    bool is_index = false;
-
-    /// if this is an integer (e.g. tqitem, point_process) variable which
-    /// is printed as array accesses
-    bool is_integer = false;
-
-    /// if the variable is qualified as constant (this is property of IndexVariable)
-    bool is_constant = false;
-
-    explicit IndexVariableInfo(std::shared_ptr<symtab::Symbol> symbol,
-                               bool is_vdata = false,
-                               bool is_index = false,
-                               bool is_integer = false)
-        : symbol(std::move(symbol))
-        , is_vdata(is_vdata)
-        , is_index(is_index)
-        , is_integer(is_integer) {}
-};
-
-
-/**
- * \class ShadowUseStatement
- * \brief Represents ion write statement during code generation
- *
- * Ion update statement needs use of shadow vectors for certain backends
- * as atomics operations are not supported on cpu backend.
- *
- * \todo If shadow_lhs is empty then we assume shadow statement not required
- */
-struct ShadowUseStatement {
-    std::string lhs;
-    std::string op;
-    std::string rhs;
-};
-
-/** \} */  // end of codegen_details
 
 
 using printer::CodePrinter;
@@ -178,7 +60,7 @@ using printer::CodePrinter;
  *    error checking. For example, see netstim.mod where we
  *    have removed return from verbatim block.
  */
-class CodegenCoreneuronCppVisitor: public visitor::ConstAstVisitor {
+class CodegenCoreneuronCppVisitor: public CodegenCppVisitor {
   protected:
     using SymbolType = std::shared_ptr<symtab::Symbol>;
 
@@ -193,25 +75,6 @@ class CodegenCoreneuronCppVisitor: public visitor::ConstAstVisitor {
      */
     using ParamVector = std::vector<std::tuple<std::string, std::string, std::string, std::string>>;
 
-    /**
-     * Code printer object for target (C++)
-     */
-    std::unique_ptr<CodePrinter> printer;
-
-    /**
-     * Name of mod file (without .mod suffix)
-     */
-    std::string mod_filename;
-
-    /**
-     * Data type of floating point variables
-     */
-    std::string float_type = codegen::naming::DEFAULT_FLOAT_TYPE;
-
-    /**
-     * Flag to indicate if visitor should avoid ion variable copies
-     */
-    bool optimize_ionvar_copies = true;
 
     /**
      * Flag to indicate if visitor should print the visited nodes
@@ -1589,10 +1452,7 @@ class CodegenCoreneuronCppVisitor: public visitor::ConstAstVisitor {
                       const std::string& output_dir,
                       std::string float_type,
                       const bool optimize_ionvar_copies)
-        : printer(std::make_unique<CodePrinter>(output_dir + "/" + mod_filename + ".cpp"))
-        , mod_filename(std::move(mod_filename))
-        , float_type(std::move(float_type))
-        , optimize_ionvar_copies(optimize_ionvar_copies) {}
+        : CodegenCppVisitor(mod_filename, output_dir, float_type, optimize_ionvar_copies) {}
 
     /**
      * \copybrief nmodl::codegen::CodegenCoreneuronCppVisitor
@@ -1614,10 +1474,7 @@ class CodegenCoreneuronCppVisitor: public visitor::ConstAstVisitor {
                       std::ostream& stream,
                       std::string float_type,
                       const bool optimize_ionvar_copies)
-        : printer(std::make_unique<CodePrinter>(stream))
-        , mod_filename(std::move(mod_filename))
-        , float_type(std::move(float_type))
-        , optimize_ionvar_copies(optimize_ionvar_copies) {}
+        : CodegenCppVisitor(mod_filename, stream, float_type, optimize_ionvar_copies) {}
 
     /**
      * Main and only member function to call after creating an instance of this class.
